requirements.txt
Django==4.2.19
django-crispy-forms==2.1
python-dotenv==1.0.0
whitenoise==6.6.0
gunicorn==21.2.0
requests==2.31.0
django-environ==0.10.0
asgiref==3.7.2
urllib3==2.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
